requirements.txt
# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
timm>=0.9.0

# FastAPI and Web Server
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
aiofiles>=23.0.0

# Optional - for training
scikit-learn>=1.3.0
matplotlib>=3.7.0
tqdm>=4.65.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
Pillow>=9.0.0
flask>=2.3.0
flask-cors>=4.0.0
numpy>=1.24.0
opencv-python>=4.8.0
matplotlib>=3.7.0
gunicorn>=21.0.0
